requirements.txt
beautifulsoup4==4.12.3
cryptography==44.0.0
detecto==1.2.2
easyocr==1.7.2
fitz==0.0.1.dev2
matplotlib==3.10.0
numpy==2.2.2
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
openpyxl==3.1.5
Pillow==11.1.0
pyautogui==0.9.54
PyPDF2==3.0.1
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
scipy==1.15.1
torch==2.5.1+cu118
ultralytics==8.3.68
xlwings==0.33.6
